integration/python_api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pandas
openpyxl
pywin32
sqlalchemy>=2.0.0
psycopg2-binary
alembic
redis
python-dotenv
sse-starlette
passlib==1.7.4
bcrypt==4.1.2
